requirement.txt
pandas
tweepy
psycopg2-binary
sqlalchemy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client 